Remove circular import of colors/text
In the education repo, we configure SASS/SCSS to add an `import` of the `variables.scss` file to the top of every file. Previously this has worked fine, but when we upgrade to a newer version of `sass`/`sass-loader` it starts breaking because it creates a circular import of the `colors/text` file.

I've chosen to resolve this by removing the `colors/text` import from `index.js` and the associated `defaultFont` theme variable. I've not been able to find any usages of the `defaultFont` variable anywhere so I'm pretty confident that variable is unused in our app.